[netstack] Add monitoring feature to ifinfo
Also its UX is enhanced.

Eg. of supported commands:

$ ifinfo --help
Usage: ifinfo [OPTIONS] [NIC_NAME]
  -period uint
        stat update period in sec (default 1)
  -sec uint
        how many sec to watch. 0 for forever (default 10)
  -watch
        watch a NIC every second

$ ifinfo -watch -sec 30 -period 5 en2
en2: Rx [   0s ago]        246 pkts      Tx [   0s ago]        240 pkts      2017-09-08 05:29:07 +0000 UTC

$ ifinfo en2

en2:
Up since: 2017-09-08 05:27:50 +0000 UTC (2m4s ago)

Bucket         Unit   Rx           Tx
-------------- ------ ------------ ------------
         Total    [#]          367          359
         Total    [B]          367          358

NET-166 #done

+++ go/src/netstack/eth/eth.go
@@ -65,16 +65,16 @@ type Client struct {
 	txDepth int
 	rxDepth int
 
-	mu         sync.Mutex
-	state      State
-	stateFunc  func(State)
-	arena      *Arena
-	tmpbuf     []bufferEntry // used to fill rx and drain tx
-	recvbuf    []bufferEntry // packets received
-	sendbuf    []bufferEntry // packets ready to send
+	mu        sync.Mutex
+	state     State
+	stateFunc func(State)
+	arena     *Arena
+	tmpbuf    []bufferEntry // used to fill rx and drain tx
+	recvbuf   []bufferEntry // packets received
+	sendbuf   []bufferEntry // packets ready to send
 
 	// Stats are informational purpose, to be reported to upper layers.
-	Stats nsfidl.NetInterfaceStats
+	Stats nsfidl.NetInterfaceStats // TODO (porce): Replace with ifState.
 
 	// These are counters for buffer management purpose.
 	txTotal    int
@@ -130,7 +130,7 @@ func NewClient(clientName, path string, arena *Arena, stateFunc func(State)) (*C
 		recvbuf:   make([]bufferEntry, 0, rxDepth),
 		sendbuf:   make([]bufferEntry, 0, txDepth),
 
-		Stats:     nsfidl.NetInterfaceStats{
+		Stats: nsfidl.NetInterfaceStats{
 			UpSince: time.Now().Unix(),
 		},
 	}
@@ -263,7 +263,7 @@ func (c *Client) txCompleteLocked() (bool, error) {
 
 	// Stats for report purpose
 	c.Stats.TxPktsTotal++
-	c.Stats.TxBytesTotal += uint32(n)
+	c.Stats.TxBytesTotal += uint64(n)
 
 	c.txInFlight -= n
 	c.txTotal += n
@@ -280,7 +280,7 @@ func (c *Client) txCompleteLocked() (bool, error) {
 func (c *Client) popRecvLocked() Buffer {
 	// Stats for informational purpose.
 	c.Stats.RxPktsTotal++
-	c.Stats.RxBytesTotal += uint32(len(c.recvbuf))
+	c.Stats.RxBytesTotal += uint64(len(c.recvbuf))
 
 	c.rxTotal++
 	b := c.recvbuf[0]
